speed-up .week for #147
