+++ scratch.js
@@ -1,4 +1,6 @@
 const spacetime = require('./src/index')
 
-let s = spacetime('may 30 2019', 'toksyo')
-console.log(s.isValid())
+console.time('time')
+let s = spacetime('june 20 2019')
+s.week()
+console.timeEnd('time')
